#6001 feat(refactor): match sub-component structure with rest of app; remove unused babel-plugin
