+++ babel.config.js
@@ -8,8 +8,7 @@ module.exports = function (api) {
   ];
 
   const plugins = [
-    '@babel/plugin-proposal-optional-chaining',
-    '@babel/plugin-transform-runtime'
+    '@babel/plugin-proposal-optional-chaining'
   ];
 
   return {
