+++ src/components/NewsletterForm/Consent/NewsletterFormConsent.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import cx from 'classnames';
 
-import NewsletterFormItem from 'NewsletterForm/Item';
+import NewsletterFormItem from '../Item/NewsletterFormItem';
 
 type NewsletterFormConsentProps = {
   checked?: boolean | null;
