--- src/components/NewsletterForm/Consent/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NewsletterFormConsent';
