--- src/components/NewsletterForm/Email/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NewsletterFormEmail';
