--- src/components/NewsletterForm/Item/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NewsletterFormItem';
