--- src/components/NewsletterForm/Submit/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NewsletterFormSubmit';
